Working ajax page displaying a fake read from distance thread

+++ avr-uip/apps/simple_ajax/simple_ajax.h
@@ -35,7 +35,7 @@ struct simple_httpd_state {
 };
 
 
-extern int range;
+extern uint16_t range;
 
 
 void simple_httpd_appcall(void);
